--- src/main/java/com/raggle/half_dream/mixin/TargetGoalMixin.java
@@ -1,10 +1,0 @@
-package com.raggle.half_dream.mixin;
-
-import org.spongepowered.asm.mixin.Mixin;
-
-import net.minecraft.entity.ai.goal.TargetGoal;
-
-@Mixin(TargetGoal.class)
-public class TargetGoalMixin {
-
-}
