+++ src/main/java/com/raggle/half_dream/mixin/client/ClientPlayerEntityMixin.java
@@ -1,4 +1,4 @@
-package com.raggle.half_dream.mixin;
+package com.raggle.half_dream.mixin.client;
 
 import org.spongepowered.asm.mixin.Mixin;
