+++ src/main/resources/half_dream.compat.mixins.json
@@ -1,0 +1,14 @@
+{
+	"required": true,
+	"minVersion": "0.8",
+	"package": "com.raggle.half_dream.compat.mixin",
+	"compatibilityLevel": "JAVA_17",
+	"mixins": [
+	],
+	"client": [
+		"sodium/WorldSliceMixin"
+	],
+	"injectors": {
+		"defaultRequire": 1
+	}
+}
