+++ src/main/resources/half_dream.mixins.common.json
@@ -1,0 +1,26 @@
+{
+	"required": true,
+	"minVersion": "0.8",
+	"package": "com.raggle.half_dream.mixin.common",
+	"compatibilityLevel": "JAVA_17",
+	"mixins": [
+		"AbstractBlockStateMixin",
+		"AoCalculatorMixin",
+		"BrewingRecipeRegistryMixin",
+		"BlockMixin",
+		"EntityMixin",
+		"ItemEntityMixin",
+		"ItemScattererMixin",
+		"LandPathNodeMakerMixin",
+		"LivingEntityMixin",
+		"MobEntityAccessor",
+		"MobEntityMixin",
+		"PlayerEntityMixin",
+		"ServerPlayerEntityMixin",
+		"SkeletonHorseEntityMixin",
+		"WorldMixin"
+	],
+	"injectors": {
+		"defaultRequire": 1
+	}
+}
